server/agents/security/requirements.txt
fastapi==0.104.1
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
uvicorn==0.24.0
semgrep==1.52.0
trivy==0.48.1
dependency-check==8.4.0